+++ circutBoardWorkFlow/Controllers/CircuitBoardController.cs
@@ -18,10 +18,10 @@ namespace circutBoardWorkFlow.Controllers
         public async Task<Result<CircuitBoardViewModel>> Create(CircuitBoardViewModel newBoard) => await _circuitBoardService.Create(newBoard);
 
         [HttpPatch, Route("/update/{boardId}")]
-        public async Task<Result<CircuitBoardViewModel>> Update(uint boardId, [FromBody]CircuitBoardViewModel board) => await _circuitBoardService.Update(boardId, board);
+        public async Task<Result<CircuitBoardViewModel>> Update(long boardId, [FromBody]CircuitBoardViewModel board) => await _circuitBoardService.Update(boardId, board);
 
         [HttpGet, Route("/history/{boardId}")]
-        public async Task<Result<ICollection<HistoryRecordViewModel>>> GetHistory(uint boardId) => await _circuitBoardService.GetHistory(boardId);
+        public async Task<Result<ICollection<HistoryRecordViewModel>>> GetHistory(long boardId) => await _circuitBoardService.GetHistory(boardId);
 
     }
 }
